Modify qemu and test script to better regress
For qemu we are using security_model=none now.
For test script we are attempting to make sure
we are using a exported file system versus something
weird like tmp or macosx.  We may have permissions
problems with creation of test directory on workers
so this may be something we have to modify as right
now it is hardcoded to /workspaces/tmp

Signed-off-by: Eric Van Hensbergen <ericvh@kernel.org>